Break down main spec file
