+++ test/runner.js
@@ -11,7 +11,14 @@ var mocha = new Mocha();
  */
 mocha.timeout(1200000);
 mocha.reporter('spec');
-mocha.addFile(__dirname + '/main.spec.js');
+mocha.addFile(__dirname + '/plugins/curly-quotes-plugin.spec.js');
+mocha.addFile(__dirname + '/plugins/smart-lists-plugin.spec.js');
+mocha.addFile(__dirname + '/plugins/toolbar-plugin.spec.js');
+mocha.addFile(__dirname + '/block-mode.spec.js');
+mocha.addFile(__dirname + '/commands.spec.js');
+mocha.addFile(__dirname + '/formatters.spec.js');
+mocha.addFile(__dirname + '/inline-elements-mode.spec.js');
+mocha.addFile(__dirname + '/patches.spec.js');
 mocha.addFile(__dirname + '/undo-manager.spec.js');
 
 createRunner(mocha);
